Implemented light & dark mode.

+++ lib/screens/task/add_task_screen.dart
@@ -6,8 +6,9 @@ import 'package:intl/intl.dart';
 
 import '../../main.dart';
 import '../../models/task.dart';
-import '../../widgets/dialogs.dart';
+import '../../utils/dialogs.dart';
 import '../../widgets/custom_banner_ad.dart';
+import '../../widgets/custom_text_form_field.dart';
 
 class AddTaskScreen extends StatefulWidget {
   const AddTaskScreen({super.key, required this.task});
@@ -138,7 +139,6 @@ class _AddTaskScreenState extends State<AddTaskScreen> {
             CustomTextFormField(
               controller: _descriptionController,
               hintText: 'Add note',
-              isForDescription: true,
               onFieldSubmitted: (value) {
                 _descriptionController.text = value;
               },
@@ -158,22 +158,62 @@ class _AddTaskScreenState extends State<AddTaskScreen> {
                 if (_isTaskAlreadyExist())
                   ElevatedButton.icon(
                       onPressed: () {
-                        _deleteTask();
-                        Dialogs.showSnackBar(
-                            context, 'Task deleted successfully!');
-                        Navigator.pop(context);
+                        showDialog(
+                          context: context,
+                          builder: (ctx) => AlertDialog(
+                            title: const Text('Are you sure?'),
+                            content: const Text('Do you want to delete this?'),
+                            actions: <Widget>[
+                              Row(
+                                  mainAxisAlignment:
+                                      MainAxisAlignment.spaceAround,
+                                  children: [
+                                    TextButton(
+                                        child: const Text(
+                                          'Yes',
+                                          style: TextStyle(color: Colors.red),
+                                        ),
+                                        onPressed: () {
+                                          _deleteTask();
+                                          Dialogs.showSnackBar(context,
+                                              'Task deleted successfully!');
+                                          Navigator.pop(context);
+                                          Navigator.pop(context);
+                                        }),
+                                    TextButton(
+                                        child: Text(
+                                          'No',
+                                          style: TextStyle(
+                                              color: Theme.of(context)
+                                                  .colorScheme
+                                                  .secondary),
+                                        ),
+                                        onPressed: () {
+                                          Navigator.of(ctx).pop(false);
+                                        }),
+                                  ])
+                            ],
+                          ),
+                        );
                       },
-                      style: const ButtonStyle(
-                          backgroundColor: MaterialStatePropertyAll(Colors.red),
-                          foregroundColor:
-                              MaterialStatePropertyAll(Colors.white)),
+                      style: ElevatedButton.styleFrom(
+                        alignment: Alignment.center,
+                        backgroundColor: Colors.red,
+                        foregroundColor: Colors.white,
+                        shape: RoundedRectangleBorder(
+                            borderRadius: BorderRadius.circular(8)),
+                      ),
                       icon: const Icon(CupertinoIcons.delete),
                       label: const Text('Delete')),
                 ElevatedButton.icon(
                   onPressed: () => _isTaskAlreadyExistUpdateElseCreate(),
                   style: ElevatedButton.styleFrom(
-                      foregroundColor: Colors.white,
-                      backgroundColor: Colors.blue),
+                    alignment: Alignment.center,
+                    backgroundColor: Colors.blue,
+                    foregroundColor: Colors.white,
+                    shape: RoundedRectangleBorder(
+                        borderRadius: BorderRadius.circular(8)),
+                  ),
                   icon: Icon(_isTaskAlreadyExist()
                       ? CupertinoIcons.refresh_thick
                       : CupertinoIcons.list_bullet_indent),
@@ -244,10 +284,10 @@ class _AddTaskScreenState extends State<AddTaskScreen> {
                           ? _time!
                           : DateFormat.yMMMEd().format(_date!),
                       style: const TextStyle(
-                          fontWeight: FontWeight.bold,
-                          fontSize: 18,
-                          letterSpacing: 1,
-                          color: Colors.black)),
+                        fontWeight: FontWeight.bold,
+                        fontSize: 18,
+                        letterSpacing: 1,
+                      )),
                 ],
               ),
             ),
@@ -256,8 +296,8 @@ class _AddTaskScreenState extends State<AddTaskScreen> {
               width: 80,
               height: 35,
               decoration: BoxDecoration(
-                color: Colors.grey.shade200,
                 borderRadius: BorderRadius.circular(10),
+                color: Theme.of(context).colorScheme.primaryContainer,
               ),
               child: Center(
                 child: Text(dateTime,
@@ -265,50 +305,6 @@ class _AddTaskScreenState extends State<AddTaskScreen> {
               ),
             ),
           ],
-        ),
-      ),
-    );
-  }
-}
-
-class CustomTextFormField extends StatelessWidget {
-  const CustomTextFormField({
-    super.key,
-    required this.controller,
-    required this.hintText,
-    required this.onFieldSubmitted,
-    this.isForDescription = false,
-  });
-
-  final TextEditingController? controller;
-  final Function(String)? onFieldSubmitted;
-
-  final String hintText;
-  final bool isForDescription;
-
-  @override
-  Widget build(BuildContext context) {
-    return TextFormField(
-      controller: controller,
-      onFieldSubmitted: onFieldSubmitted,
-      cursorColor: Colors.blue,
-      style: const TextStyle(
-          fontWeight: FontWeight.bold, fontSize: 18, letterSpacing: 1),
-      decoration: InputDecoration(
-        prefixIcon: isForDescription
-            ? const Icon(Icons.bookmark_border_rounded, color: Colors.grey)
-            : const Icon(Icons.sports_gymnastics_rounded, color: Colors.grey),
-        hintText: hintText,
-        hintStyle:
-            const TextStyle(fontWeight: FontWeight.bold, color: Colors.grey),
-        enabledBorder: OutlineInputBorder(
-          borderRadius: BorderRadius.circular(10),
-          borderSide: BorderSide(
-              color: Theme.of(context).colorScheme.secondary.withOpacity(.4)),
-        ),
-        focusedBorder: OutlineInputBorder(
-          borderRadius: BorderRadius.circular(10),
-          borderSide: const BorderSide(color: Colors.lightBlue),
         ),
       ),
     );
